Change http transport interface
